feat: optionally add human-readable labels. (#5)
OxQL provides uuid labels for silos, projects, etc., but these aren't very
useful for metrics queries: the user has to look up the uuid for each
silo/project of interest. This patch adds the option to fetch silo and project
metadata from the api and enrich labels where appropriate. We should support
this in the api instead eventually, but it's quicker to add here for now.